firebuild: Squeeze ack_id to 16 bits

+++ src/firebuild/utils.h
@@ -38,7 +38,7 @@ ssize_t fb_copy_file_range(int fd_in, loff_t *off_in, int fd_out, loff_t *off_ou
 
 namespace firebuild {
 
-void ack_msg(const int conn, const uint32_t ack_num);
+void ack_msg(const int conn, const uint16_t ack_num);
 
 std::string make_fifo(int fd, int flags, int pid, const char *fb_conn_string,
                       int *fifo_name_offset);
